MLS utilizing data federation via blockchain
This code sample demonstrates a simple implementation of multi-level security and data federation through blockchain. The contract includes three levels of access: admin, federation, and user. Admins have the highest level of access and can add or remove other admins, as well as add or remove federations. Federations have the ability to add or remove users, and users can add data.

The data is stored in a mapping, with each data point assigned a unique ID. Federations have access to the data and can retrieve it using the ID. Users can only add data and cannot retrieve it.

This implementation is just a basic example and can be expanded upon to include more features and functionality. For example, additional levels of access could be added, or the data storage system could be made more complex to include more data types and structures.